refactor: move project menu to dedicated component
The popup menu on the project page has been broken out into a dedicated
ProjectMenu component to reduce redundancy between the two project
page routes.
